Separate component for navigaition buttons
