+++ src/content/app/tools/blast/components/blast-navigation-buttons/BlastNavigationButtons.scss
@@ -1,0 +1,5 @@
+.actionButtons {
+  display: flex;
+  column-gap: 12px;
+  align-self: start;
+}
