+++ src/content/app/tools/blast/components/blast-settings/BlastSettings.scss
@@ -24,12 +24,6 @@
   margin-right: 35px;
 }
 
-.actionButtons {
-  display: flex;
-  column-gap: 12px;
-  align-self: start;
-}
-
 .select label {
   display: grid;
   grid-template-columns: repeat(2, auto);
